Added handlers for NICK and USERS for newly connected users. (fixes #8)
Also party added PING and WHOIS.  WHOIS is currenly only handling
simple username lookups, need to parse the query and do some more
fancy searching in order to be fully compliant with the spec.

PING messages are currently only answered with a PONG, the server
never sends any PING at this stage.  In order for this to work the
Connection must either have a timeout for when to stop handling
incoming messages and send a PING, or the server needs to handle
this somewhere else.

There is currently no support in joingning channels yet, and no
support for SENDPRIVMSG has been added.  Should also create a
message class for handling message codes &c.